fix zoom bug
The Library "barteksc - AndroidPdfViewer" had a zoom limit of 10x (variable MAXIMUM_ZOOM in class "Pinch.java" in said Library) which did not allow to pinch over that value (methid "onScale" in class "DragPinchManager.java"). Nonethless, it was possible to double-tap-to-zoom to any value, even above 10x, causing a poor UX.

For this reason, the maximum zoom has been set at 10x with a SeekBar, instead of using an EditText which allowed any value.